Generalise and improve tests
check_magma and other functions now do not rely on Boost.Test anymore.
That dependency has been moved to a separate file.
The functions also make it possible to switch checks on and off.
They also can generate a report.
This is useful for a Python wrapper of semirings, say.